restructuration et vérification des fonctionnalités
