+++ navigation/localisation/src/LaserScan.cpp
