+++ navigation/localisation/src/Segment.cpp
@@ -2,6 +2,7 @@
 #include "Model.h"
 #include "Segment.h"
 #include "cartographie_utils.h"
+#include "math_functions.h"
 
 Segment::Segment() : m_angle(0.0),m_size(0.0)
 {
@@ -32,6 +33,15 @@ geometry_msgs::Point Segment::getMax()
 	return m_max;
 }
 
+geometry_msgs::Point Segment::getCenter()
+{
+	geometry_msgs::Point middle;
+	middle.x = (getMax().x + getMin().x)/2;
+	middle.y = (getMax().y + getMin().y)/2;
+
+	return middle;
+}
+
 void Segment::setAngle(double theta)
 {
 	m_angle = atan2(tan(theta),1);
@@ -59,6 +69,8 @@ void Segment::setPoints(geometry_msgs::Point a, geometry_msgs::Point b)
 
 void Segment::update()
 {
-	m_angle = tan((m_max.y-m_min.y)/(m_max.x-m_min.x));
+	m_angle = atan((m_max.y-m_min.y)/(m_max.x-m_min.x));
 	m_size  = dist(m_min,m_max);
+	m_max.z = m_angle;
+	m_min.z = m_angle;
 }
