+++ navigation/localisation/src/fake_scan_node.cpp
@@ -21,10 +21,9 @@ int main(int argc, char** argv)
     scan->range_min = 0;
     scan->range_max = 5;
 
-    std::vector<float> ranges = {0.9679999947547913, 0.9679999947547913, 0.9679999947547913, 0.9700000286102295, 0.9729999899864197, 0.9779999852180481, 0.9779999852180481, 0.9810000061988831, 0.9810000061988831, 0.9710000157356262, 0.9710000157356262, 0.9710000157356262, 0.9710000157356262, 0.9670000076293945, 0.9700000286102295, 0.9700000286102295, 0.9670000076293945, 0.9670000076293945, 0.9660000205039978, 0.9660000205039978, 0.9660000205039978, 0.9750000238418579, 0.9769999980926514, 0.9769999980926514, 0.9750000238418579, 0.9769999980926514, 0.9819999933242798, 0.9819999933242798, 0.9829999804496765, 0.984000027179718, 0.9919999837875366, 1.0, 0.9919999837875366, 0.9919999837875366, 0.9940000176429749, 0.9940000176429749, 0.9940000176429749, 0.9959999918937683, 0.996999979019165, 1.003999948501587, 0.996999979019165, 0.9959999918937683, 0.9959999918937683, 0.9959999918937683, 0.9959999918937683, 0.996999979019165, 0.9980000257492065, 1.0089999437332153, 1.0089999437332153, 1.0130000114440918, 1.0089999437332153, 1.0110000371932983, 1.0110000371932983, 1.0110000371932983, 1.0110000371932983, 1.0119999647140503, 1.0169999599456787, 1.027999997138977, 1.027999997138977, 1.027999997138977, 1.027999997138977, 1.027999997138977, 1.0299999713897705, 1.034999966621399, 1.0449999570846558, 1.0449999570846558, 1.0449999570846558, 1.0460000038146973, 1.0460000038146973, 1.0490000247955322, 1.0499999523162842, 1.0520000457763672, 1.0570000410079956, 1.062000036239624, 1.0640000104904175, 1.0729999542236328, 1.0750000476837158, 1.0759999752044678, 1.0789999961853027, 1.0829999446868896, 1.0829999446868896, 1.0829999446868896, 1.0850000381469727, 1.0850000381469727, 1.0920000076293945, 1.100000023841858, 1.1019999980926514, 1.1030000448226929, 1.1109999418258667, 1.1139999628067017, 1.1150000095367432, 1.1150000095367432, 1.1239999532699585, 1.1239999532699585, 1.128999948501587, 1.1319999694824219, 1.1339999437332153, 1.1449999809265137, 1.1490000486373901, 1.1490000486373901, 1.152999997138977, 1.1610000133514404, 1.1610000133514404, 1.1759999990463257, 1.180999994277954, 1.1829999685287476, 1.1890000104904175, 1.190999984741211, 1.1929999589920044, 1.1959999799728394, 1.1979999542236328, 1.215999960899353, 1.2200000286102295, 1.2269999980926514, 1.2319999933242798, 1.2319999933242798, 1.2330000400543213, 1.2380000352859497, 1.2380000352859497, 1.2450000047683716, 1.2450000047683716, 1.2289999723434448, 0.3930000066757202, 0.38999998569488525, 0.3889999985694885, 0.3889999985694885, 0.3880000114440918, 0.3709999918937683, 0.3709999918937683, 0.3709999918937683, 0.3709999918937683, 0.3709999918937683, 0.3709999918937683, 0.35899999737739563, 0.35899999737739563, 0.3580000102519989, 0.3569999933242798, 0.3569999933242798, 0.3479999899864197, 0.34700000286102295, 0.34700000286102295, 0.34599998593330383, 0.34599998593330383, 0.3440000116825104, 0.34299999475479126, 0.34200000762939453, 0.34200000762939453, 0.335999995470047, 0.33399999141693115, 0.32899999618530273, 0.32499998807907104, 0.3240000009536743, 0.3240000009536743, 0.3240000009536743, 0.3230000138282776, 0.3160000145435333, 0.3160000145435333, 0.3240000009536743, 0.3160000145435333, 0.3160000145435333, 0.3149999976158142, 0.3149999976158142, 0.3149999976158142, 0.31299999356269836, 0.3089999854564667, 0.3089999854564667, 0.30799999833106995, 0.30000001192092896, 0.30000001192092896, 0.296999990940094, 0.29600000381469727, 0.29600000381469727, 0.2939999997615814, 0.2939999997615814, 0.2939999997615814, 0.2939999997615814, 0.2939999997615814, 0.289000004529953, 0.289000004529953, 0.289000004529953, 0.2939999997615814, 0.2939999997615814, 0.2939999997615814, 0.2879999876022339, 0.28700000047683716, 0.28700000047683716, 0.28700000047683716, 0.28700000047683716, 0.28600001335144043, 0.28200000524520874, 0.2809999883174896, 0.2800000011920929, 0.27900001406669617, 0.27900001406669617, 0.27900001406669617, 0.27900001406669617, 0.27900001406669617, 0.27900001406669617, 0.27799999713897705, 0.2759999930858612, 0.2750000059604645, 0.2750000059604645, 0.27300000190734863, 0.27300000190734863, 0.26600000262260437, 0.26499998569488525, 0.26499998569488525, 0.26600000262260437, 0.26600000262260437, 0.26499998569488525, 0.2639999985694885, 0.2630000114440918, 0.2630000114440918, 0.2630000114440918, 0.2630000114440918, 0.2630000114440918, 0.2590000033378601, 0.2590000033378601, 0.2590000033378601, 0.2590000033378601, 0.26100000739097595, 0.2669999897480011, 0.26100000739097595, 0.2680000066757202, 0.2680000066757202, 0.2590000033378601, 0.257999986410141, 0.257999986410141, 0.257999986410141, 0.25699999928474426, 0.25699999928474426, 0.2590000033378601, 0.2619999945163727, 0.2590000033378601, 0.26100000739097595, 0.26100000739097595, 0.26100000739097595, 0.26100000739097595, 0.26600000262260437, 0.26600000262260437, 0.26600000262260437, 0.2590000033378601, 0.2590000033378601, 0.25600001215934753, 0.2590000033378601, 0.2619999945163727, 0.26899999380111694, 0.26899999380111694, 0.2800000011920929, 0.2809999883174896, 0.28600001335144043, 0.28600001335144043, 0.28600001335144043, 0.289000004529953, 0.289000004529953, 0.2919999957084656, 0.2919999957084656, 0.2930000126361847, 0.2930000126361847, 0.2939999997615814, 0.2939999997615814, 0.2939999997615814, 0.289000004529953, 0.289000004529953, 0.2840000092983246, 0.27300000190734863, 0.27300000190734863, 0.2720000147819519, 0.2669999897480011, 0.2669999897480011, 0.2669999897480011, 0.2669999897480011, 0.26100000739097595, 0.26100000739097595, 0.26100000739097595, 0.2639999985694885, 0.2639999985694885, 0.2639999985694885, 0.2619999945163727, 0.25999999046325684, 0.257999986410141, 0.257999986410141, 0.25600001215934753, 0.257999986410141, 0.2619999945163727, 0.2619999945163727, 0.2529999911785126, 0.26499998569488525, 0.2639999985694885, 0.2639999985694885, 0.2639999985694885, 0.2639999985694885, 0.2639999985694885, 0.25999999046325684, 0.25999999046325684, 0.25999999046325684, 0.26100000739097595, 0.25999999046325684, 0.26100000739097595, 0.2680000066757202, 0.2680000066757202, 0.2680000066757202, 0.2680000066757202, 0.26899999380111694, 0.2680000066757202, 0.2709999978542328, 0.2680000066757202, 0.2680000066757202, 0.26600000262260437, 0.26600000262260437, 0.26600000262260437, 0.26600000262260437, 0.27399998903274536, 0.2750000059604645, 0.2750000059604645, 0.2750000059604645, 0.2750000059604645, 0.2750000059604645, 0.2750000059604645, 0.2750000059604645, 0.2750000059604645, 0.2750000059604645, 0.2759999930858612, 0.2759999930858612, 0.27799999713897705, 0.27900001406669617, 0.2800000011920929, 0.2800000011920929, 0.2849999964237213, 0.2849999964237213, 0.2800000011920929, 0.2800000011920929, 0.2840000092983246, 0.2800000011920929, 0.2840000092983246, 0.2840000092983246, 0.2849999964237213, 0.289000004529953, 0.2939999997615814, 0.30000001192092896, 0.30300000309944153, 0.30300000309944153, 0.30300000309944153, 0.30000001192092896, 0.2980000078678131, 0.29499998688697815, 0.29499998688697815, 0.29499998688697815, 0.3050000071525574, 0.3100000023841858, 0.31200000643730164, 0.31200000643730164, 0.31299999356269836, 0.31299999356269836, 0.31700000166893005, 0.31700000166893005, 0.31299999356269836, 0.31200000643730164, 0.31200000643730164, 0.3100000023841858, 0.3100000023841858, 0.3190000057220459, 0.32100000977516174, 0.32100000977516174, 0.33000001311302185, 0.3330000042915344, 0.33799999952316284, 0.335999995470047, 0.3409999907016754, 0.34200000762939453, 0.34200000762939453, 0.3400000035762787, 0.34200000762939453, 0.3400000035762787, 0.3400000035762787, 0.34700000286102295, 0.3540000021457672, 0.35899999737739563, 0.3610000014305115, 0.35899999737739563, 0.35899999737739563, 0.3630000054836273, 0.36800000071525574, 0.3630000054836273, 0.36800000071525574, 0.36899998784065247, 0.3720000088214874, 0.3720000088214874, 0.38199999928474426, 0.38499999046325684, 0.38499999046325684, 0.38600000739097595, 0.38600000739097595, 0.38600000739097595, 0.38600000739097595, 0.38600000739097595, 0.3880000114440918, 0.3959999978542328, 0.4020000100135803, 0.40700000524520874, 0.4090000092983246, 0.4090000092983246, 0.414000004529953, 0.41999998688697815, 0.421999990940094, 0.4230000078678131, 0.4259999990463257, 0.4339999854564667, 0.43700000643730164, 0.4390000104904175, 0.4410000145435333, 0.44699999690055847, 0.45100000500679016, 0.460999995470047, 0.4620000123977661, 0.4650000035762787, 3.007999897003174, 3.007999897003174, 3.007999897003174, 3.005000114440918, 2.996000051498413, 2.9769999980926514, 2.9630000591278076, 2.947000026702881, 2.943000078201294, 2.941999912261963, 2.9200000762939453, 2.9030001163482666, 2.9019999504089355, 2.8910000324249268, 2.882999897003174, 2.872999906539917, 2.861999988555908, 2.8529999256134033, 2.8499999046325684, 2.8259999752044678, 2.8259999752044678, 2.8259999752044678, 2.8480000495910645, 2.880000114440918, 2.9159998893737793, 2.946000099182129, 3.005000114440918, 3.013000011444092, 3.062000036239624, 3.109999895095825, 3.111999988555908, 5.599999904632568, 5.599999904632568, 5.599999904632568, 2.7179999351501465, 2.7049999237060547, 2.696000099182129, 2.686000108718872, 2.681999921798706, 2.677999973297119, 2.674999952316284, 2.6730000972747803, 2.6679999828338623, 2.6589999198913574, 2.6480000019073486, 2.6470000743865967, 2.6419999599456787, 2.640000104904175, 2.638000011444092, 2.621999979019165, 2.617000102996826, 2.617000102996826, 2.615999937057495, 2.614000082015991, 2.5989999771118164, 2.5959999561309814, 2.5929999351501465, 2.5910000801086426, 2.5820000171661377, 2.5820000171661377, 2.5769999027252197, 2.5739998817443848, 2.566999912261963, 2.561000108718872, 2.559000015258789, 2.558000087738037, 2.553999900817871, 2.5510001182556152, 2.5490000247955322, 2.5429999828338623, 2.5420000553131104, 2.5420000553131104, 2.5420000553131104, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568};
-
+    std::vector<float> ranges = {0.5870000123977661, 0.5920000076293945, 0.5870000123977661, 0.593999981880188, 0.593999981880188, 0.593999981880188, 0.5879999995231628, 0.5960000157356262, 0.5920000076293945, 0.5879999995231628, 0.5920000076293945, 0.5920000076293945, 0.5879999995231628, 0.5879999995231628, 0.5879999995231628, 0.5879999995231628, 0.5879999995231628, 0.5899999737739563, 0.5879999995231628, 0.5889999866485596, 0.5849999785423279, 0.5870000123977661, 0.5870000123977661, 0.5889999866485596, 0.5899999737739563, 0.5899999737739563, 0.5910000205039978, 0.5910000205039978, 0.5910000205039978, 0.5820000171661377, 0.5849999785423279, 0.5839999914169312, 0.5839999914169312, 0.5820000171661377, 0.5820000171661377, 0.5820000171661377, 0.5770000219345093, 0.5820000171661377, 0.5759999752044678, 0.5839999914169312, 0.5849999785423279, 0.5849999785423279, 0.5849999785423279, 0.5870000123977661, 0.5879999995231628, 0.5889999866485596, 0.5960000157356262, 0.5960000157356262, 0.5960000157356262, 0.5960000157356262, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 1.1779999732971191, 1.1720000505447388, 1.1649999618530273, 1.1640000343322754, 1.1640000343322754, 1.1649999618530273, 1.1640000343322754, 1.1640000343322754, 1.1629999876022339, 1.1610000133514404, 1.1610000133514404, 1.1610000133514404, 1.1579999923706055, 1.156000018119812, 1.1419999599456787, 1.1180000305175781, 1.1139999628067017, 1.090999960899353, 1.0820000171661377, 1.055999994277954, 0.8410000205039978, 0.8410000205039978, 0.8370000123977661, 0.8410000205039978, 0.8460000157356262, 0.847000002861023, 0.8460000157356262, 0.847000002861023, 0.8560000061988831, 0.8560000061988831, 0.8569999933242798, 0.8579999804496765, 0.8619999885559082, 0.8619999885559082, 0.8629999756813049, 0.8669999837875366, 0.8700000047683716, 0.8759999871253967, 0.878000020980835, 0.8820000290870667, 0.8820000290870667, 0.8820000290870667, 0.8939999938011169, 0.9020000100135803, 0.9039999842643738, 0.9039999842643738, 0.9049999713897705, 0.9100000262260437, 0.9129999876022339, 0.9169999957084656, 0.9179999828338623, 0.9210000038146973, 0.9380000233650208, 0.9380000233650208, 0.9390000104904175, 0.9390000104904175, 0.9399999976158142, 0.9399999976158142, 0.9490000009536743, 0.9520000219345093, 0.953000009059906, 0.9639999866485596, 0.9750000238418579, 0.9819999933242798, 0.984000027179718, 0.9850000143051147, 0.9900000095367432, 0.9909999966621399, 0.996999979019165, 1.0019999742507935, 1.0160000324249268, 1.0180000066757202, 1.0199999809265137, 1.0240000486373901, 1.0240000486373901, 1.031999945640564, 1.0410000085830688, 1.0420000553131104, 1.0520000457763672, 1.062000036239624, 1.0640000104904175, 1.0759999752044678, 1.0809999704360962, 1.0870000123977661, 1.090999960899353, 1.0989999771118164, 1.1039999723434448, 1.1089999675750732, 1.1130000352859497, 1.128000020980835, 1.1319999694824219, 1.1399999856948853, 1.156000018119812, 1.1610000133514404, 1.1670000553131104, 1.1799999475479126, 1.1890000104904175, 1.1929999589920044, 1.1979999542236328, 1.2120000123977661, 1.2130000591278076, 1.2339999675750732, 1.2430000305175781, 1.2510000467300415, 1.2640000581741333, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 2.503000020980835, 2.4670000076293945, 2.447999954223633, 2.437000036239624, 2.4170000553131104, 2.4049999713897705, 2.38100004196167, 2.3589999675750732, 2.3399999141693115, 2.3239998817443848, 2.2980000972747803, 2.2899999618530273, 2.2720000743865967, 2.25600004196167, 2.236999988555908, 2.2269999980926514, 2.2149999141693115, 2.193000078201294, 2.188999891281128, 2.1630001068115234, 2.1519999504089355, 2.140000104904175, 2.121000051498413, 2.109999895095825, 2.0910000801086426, 2.0810000896453857, 2.062999963760376, 2.058000087738037, 2.0369999408721924, 2.0329999923706055, 2.01200008392334, 2.01200008392334, 2.01200008392334, 2.433000087738037, 2.434999942779541, 5.599999904632568, 2.434999942779541, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 2.3410000801086426, 2.328000068664551, 2.322999954223633, 2.296999931335449, 2.296999931335449, 2.296999931335449, 2.296999931335449, 4.993000030517578, 5.015999794006348, 5.076000213623047, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 4.086999893188477, 4.074999809265137, 3.9779999256134033, 4.074999809265137, 4.086999893188477, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 1.6740000247955322, 1.659999966621399, 1.659000039100647, 1.659000039100647, 1.659999966621399, 1.659000039100647, 1.659000039100647, 1.659999966621399, 1.659999966621399, 1.659000039100647, 1.659000039100647, 1.659000039100647, 1.6549999713897705, 1.6519999504089355, 1.6519999504089355, 1.6519999504089355, 1.652999997138977, 1.652999997138977, 1.6540000438690186, 1.656000018119812, 1.656000018119812, 1.656000018119812, 1.656000018119812, 1.6549999713897705, 1.6549999713897705, 1.6549999713897705, 1.6549999713897705, 1.6579999923706055, 1.6699999570846558, 1.6640000343322754, 1.6640000343322754, 1.6640000343322754, 1.6649999618530273, 1.6640000343322754, 1.656000018119812, 1.656000018119812, 1.652999997138977, 1.6519999504089355, 1.6519999504089355, 1.649999976158142, 1.649999976158142, 1.6519999504089355, 1.649999976158142, 1.6519999504089355, 1.6519999504089355, 1.6549999713897705, 1.6519999504089355, 1.656000018119812, 1.656000018119812, 1.656000018119812, 1.656000018119812, 1.659999966621399, 1.6640000343322754, 1.6679999828338623, 1.6679999828338623, 1.6670000553131104, 1.6679999828338623, 1.6690000295639038, 1.6699999570846558, 1.6699999570846558, 1.6699999570846558, 1.6729999780654907, 1.6770000457763672, 1.680999994277954, 1.680999994277954, 1.684000015258789, 1.684999942779541, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 5.599999904632568, 3.140000104904175, 3.13100004196167, 3.128999948501587, 3.121999979019165, 3.11899995803833, 3.0999999046325684, 3.0999999046325684, 3.0940001010894775, 3.0789999961853027, 3.0789999961853027, 3.069999933242798, 3.069000005722046, 3.058000087738037, 3.052000045776367, 3.0510001182556152, 3.0480000972747803, 3.0409998893737793, 3.0390000343322754, 3.0360000133514404, 3.0269999504089355, 3.0269999504089355, 3.0260000228881836, 3.0160000324249268, 3.009999990463257, 3.009999990463257, 3.003000020980835, 2.996000051498413, 2.992000102996826, 2.990000009536743, 2.9860000610351562, 2.984999895095825, 2.9809999465942383, 2.9800000190734863, 2.9790000915527344, 0.7020000219345093, 0.7020000219345093, 0.7020000219345093, 0.7080000042915344, 0.7080000042915344, 0.7170000076293945, 0.7170000076293945, 0.7170000076293945, 0.7099999785423279, 0.7099999785423279, 0.7099999785423279, 0.9120000004768372, 1.1200000047683716, 1.2910000085830688, 1.2910000085830688, 1.5, 1.621000051498413, 1.7230000495910645, 1.7690000534057617, 1.7710000276565552, 1.7730000019073486, 1.7979999780654907, 1.8020000457763672, 1.805999994277954, 1.805999994277954, 2.8919999599456787, 5.599999904632568};
     scan->ranges = ranges;
-    scan->header.frame_id = "map";
+    scan->header.frame_id = "laser_link";
 
 	ros::Rate loop_rate (10);
 	while(n.ok())
